+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added offtarget module to count reads that map to off target loci in hg38 (#143).
 - Added typing rate and mapping rate information per marker to the main pipe HTML report (#146).
 - Added marker detail HTML report (#146).
+- Implemented support for single-end reads in the end-to-end microhap analysis pipeline (#147).
 
 ### Changed
 - Exposed static and dynamic threshold configuration to pipe CLI (#135).